+++ examples/tables.js
@@ -6,7 +6,8 @@ import {
     SIMPLE_TABLE,
     SIMPLE_TABLE_WITH_FOOTER,
     SIMPLE_TABLE_EXPANDABLE,
-    SIMPLE_TABLE_SORTABLE_EXPANDABLE
+    SIMPLE_TABLE_SORTABLE_EXPANDABLE,
+    SIMPLE_TABLE_SORTABLE_PAGINATED
 } from './tables.data';
 
 require('./example.less');
@@ -67,11 +68,21 @@ render(
                 <h2 className="ffe-h2">Sortable table</h2>
                 <ResponsiveTable columns={SIMPLE_TABLE.columns} data={SIMPLE_TABLE.dataObject} sortable={true}/>
 
-                <h2 className="ffe-h2">Sortable, expandable, screen reader caption, footerized, condensed, small headers</h2>
+                <h2 className="ffe-h2">Sortable table with offset and limit (5th to 15th item) and screen-reader-only caption</h2>
+                <ResponsiveTable
+                    columns={SIMPLE_TABLE_SORTABLE_PAGINATED.columns}
+                    data={SIMPLE_TABLE_SORTABLE_PAGINATED.dataObject}
+                    sortable={true}
+                    offset={4}
+                    limit={10}
+                    caption="FFE colors"
+                    srOnlyCaption={true}
+                />
+
+                <h2 className="ffe-h2">Sortable, expandable, captioned, footerized, condensed, small headers</h2>
                 <ResponsiveTable
                     columns={SIMPLE_TABLE_SORTABLE_EXPANDABLE.columns}
                     data={SIMPLE_TABLE_SORTABLE_EXPANDABLE.dataObject}
-                    srOnlyCaption={true}
                     expandedContentMapper={ lazyExpandedContentMapper }
                     condensed={true}
                     smallHeader={true}